requirements.txt
numpy
transformers==4.10.0
tensorboard
nltk
rouge_score
torchmetrics
wandb
scipy
sentencepiece
pytorch-metric-learning
prettytable
fairscale
